+++ packages/providers/lib.esm/etherscan-provider.js
@@ -17,19 +17,20 @@ import { version } from "./_version";
 const logger = new Logger(version);
 import { BaseProvider } from "./base-provider";
 // The transaction has already been sanitized by the calls in Provider
-function getTransactionString(transaction) {
-    const result = [];
+function getTransactionPostData(transaction) {
+    const result = {};
     for (let key in transaction) {
         if (transaction[key] == null) {
             continue;
         }
         let value = hexlify(transaction[key]);
+        // Quantity-types require no leading zero, unless 0
         if ({ gasLimit: true, gasPrice: true, nonce: true, value: true }[key]) {
             value = hexValue(value);
         }
-        result.push(key + "=" + value);
+        result[key] = value;
     }
-    return result.join("&");
+    return result;
 }
 function getResult(result) {
     // getLogs, getHistory have weird success responses
@@ -165,12 +166,12 @@ export class EtherscanProvider extends BaseProvider {
             perform: { get: () => super.perform }
         });
         return __awaiter(this, void 0, void 0, function* () {
-            let url = this.baseUrl;
+            let url = this.baseUrl + "/api";
             let apiKey = "";
             if (this.apiKey) {
                 apiKey += "&apikey=" + this.apiKey;
             }
-            const get = (url, procFunc) => __awaiter(this, void 0, void 0, function* () {
+            const get = (url, payload, procFunc) => __awaiter(this, void 0, void 0, function* () {
                 this.emit("debug", {
                     action: "request",
                     request: url,
@@ -186,7 +187,14 @@ export class EtherscanProvider extends BaseProvider {
                         return Promise.resolve(true);
                     }
                 };
-                const result = yield fetchJson(connection, null, procFunc || getJsonResult);
+                let payloadStr = null;
+                if (payload) {
+                    connection.headers = { "content-type": "application/x-www-form-urlencoded; charset=UTF-8" };
+                    payloadStr = Object.keys(payload).map((key) => {
+                        return `${key}=${payload[key]}`;
+                    }).join("&");
+                }
+                const result = yield fetchJson(connection, payloadStr, procFunc || getJsonResult);
                 this.emit("debug", {
                     action: "response",
                     request: url,
@@ -197,38 +205,41 @@ export class EtherscanProvider extends BaseProvider {
             });
             switch (method) {
                 case "getBlockNumber":
-                    url += "/api?module=proxy&action=eth_blockNumber" + apiKey;
-                    return get(url);
+                    url += "?module=proxy&action=eth_blockNumber" + apiKey;
+                    return get(url, null);
                 case "getGasPrice":
-                    url += "/api?module=proxy&action=eth_gasPrice" + apiKey;
-                    return get(url);
+                    url += "?module=proxy&action=eth_gasPrice" + apiKey;
+                    return get(url, null);
                 case "getBalance":
                     // Returns base-10 result
-                    url += "/api?module=account&action=balance&address=" + params.address;
+                    url += "?module=account&action=balance&address=" + params.address;
                     url += "&tag=" + params.blockTag + apiKey;
-                    return get(url, getResult);
+                    return get(url, null, getResult);
                 case "getTransactionCount":
-                    url += "/api?module=proxy&action=eth_getTransactionCount&address=" + params.address;
+                    url += "?module=proxy&action=eth_getTransactionCount&address=" + params.address;
                     url += "&tag=" + params.blockTag + apiKey;
-                    return get(url);
+                    return get(url, null);
                 case "getCode":
-                    url += "/api?module=proxy&action=eth_getCode&address=" + params.address;
+                    url += "?module=proxy&action=eth_getCode&address=" + params.address;
                     url += "&tag=" + params.blockTag + apiKey;
-                    return get(url);
+                    return get(url, null);
                 case "getStorageAt":
-                    url += "/api?module=proxy&action=eth_getStorageAt&address=" + params.address;
+                    url += "?module=proxy&action=eth_getStorageAt&address=" + params.address;
                     url += "&position=" + params.position;
                     url += "&tag=" + params.blockTag + apiKey;
-                    return get(url);
+                    return get(url, null);
                 case "sendTransaction":
-                    url += "/api?module=proxy&action=eth_sendRawTransaction&hex=" + params.signedTransaction;
-                    url += apiKey;
-                    return get(url).catch((error) => {
+                    return get(url, {
+                        module: "proxy",
+                        action: "eth_sendRawTransaction",
+                        hex: params.signedTransaction,
+                        apikey: this.apiKey
+                    }).catch((error) => {
                         return checkError("sendTransaction", error, params.signedTransaction);
                     });
                 case "getBlock":
                     if (params.blockTag) {
-                        url += "/api?module=proxy&action=eth_getBlockByNumber&tag=" + params.blockTag;
+                        url += "?module=proxy&action=eth_getBlockByNumber&tag=" + params.blockTag;
                         if (params.includeTransactions) {
                             url += "&boolean=true";
                         }
@@ -236,51 +247,46 @@ export class EtherscanProvider extends BaseProvider {
                             url += "&boolean=false";
                         }
                         url += apiKey;
-                        return get(url);
+                        return get(url, null);
                     }
                     throw new Error("getBlock by blockHash not implemented");
                 case "getTransaction":
-                    url += "/api?module=proxy&action=eth_getTransactionByHash&txhash=" + params.transactionHash;
+                    url += "?module=proxy&action=eth_getTransactionByHash&txhash=" + params.transactionHash;
                     url += apiKey;
-                    return get(url);
+                    return get(url, null);
                 case "getTransactionReceipt":
-                    url += "/api?module=proxy&action=eth_getTransactionReceipt&txhash=" + params.transactionHash;
+                    url += "?module=proxy&action=eth_getTransactionReceipt&txhash=" + params.transactionHash;
                     url += apiKey;
-                    return get(url);
+                    return get(url, null);
                 case "call": {
-                    let transaction = getTransactionString(params.transaction);
-                    if (transaction) {
-                        transaction = "&" + transaction;
-                    }
-                    url += "/api?module=proxy&action=eth_call" + transaction;
-                    //url += "&tag=" + params.blockTag + apiKey;
                     if (params.blockTag !== "latest") {
                         throw new Error("EtherscanProvider does not support blockTag for call");
                     }
-                    url += apiKey;
+                    const postData = getTransactionPostData(params.transaction);
+                    postData.module = "proxy";
+                    postData.action = "eth_call";
+                    postData.apikey = this.apiKey;
                     try {
-                        return yield get(url);
+                        return yield get(url, postData);
                     }
                     catch (error) {
                         return checkError("call", error, params.transaction);
                     }
                 }
                 case "estimateGas": {
-                    let transaction = getTransactionString(params.transaction);
-                    if (transaction) {
-                        transaction = "&" + transaction;
-                    }
-                    url += "/api?module=proxy&action=eth_estimateGas&" + transaction;
-                    url += apiKey;
+                    const postData = getTransactionPostData(params.transaction);
+                    postData.module = "proxy";
+                    postData.action = "eth_estimateGas";
+                    postData.apikey = this.apiKey;
                     try {
-                        return yield get(url);
+                        return yield get(url, postData);
                     }
                     catch (error) {
                         return checkError("estimateGas", error, params.transaction);
                     }
                 }
                 case "getLogs": {
-                    url += "/api?module=logs&action=getLogs";
+                    url += "?module=logs&action=getLogs";
                     if (params.filter.fromBlock) {
                         url += "&fromBlock=" + checkLogTag(params.filter.fromBlock);
                     }
@@ -304,7 +310,7 @@ export class EtherscanProvider extends BaseProvider {
                         }
                     }
                     url += apiKey;
-                    const logs = yield get(url, getResult);
+                    const logs = yield get(url, null, getResult);
                     // Cache txHash => blockHash
                     let txs = {};
                     // Add any missing blockHash to the logs
@@ -327,9 +333,9 @@ export class EtherscanProvider extends BaseProvider {
                     if (this.network.name !== "homestead") {
                         return 0.0;
                     }
-                    url += "/api?module=stats&action=ethprice";
+                    url += "?module=stats&action=ethprice";
                     url += apiKey;
-                    return parseFloat((yield get(url, getResult)).ethusd);
+                    return parseFloat((yield get(url, null, getResult)).ethusd);
                 default:
                     break;
             }