+++ packages/wallet/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "wallet/5.0.6";
+export declare const version = "wallet/5.0.7";